refactor(cli): move undo/redo tests into their own module
